refactor: move the topology components under topology
