+++ src/components/Topology/Sidebar/TopologyDetails.tsx
@@ -9,8 +9,8 @@ import { Age } from "../../../ui/Age";
 import CollapsiblePanel from "../../CollapsiblePanel";
 import ConfigCostValue from "../../Configs/ConfigCosts/ConfigCostValue";
 import Title from "../../Title/title";
-import { CardMetrics } from "../../TopologyCard/CardMetrics";
-import { TopologyLink } from "../../TopologyLink";
+import { CardMetrics } from "../TopologyCard/CardMetrics";
+import { TopologyLink } from "../TopologyLink";
 import DisplayDetailsRow from "../../Utils/DisplayDetailsRow";
 import { DisplayGroupedProperties } from "../../Utils/DisplayGroupedProperties";
 import { formatTopologyProperties } from "./Utils/formatTopologyProperties";
