+++ src/components/Topology/Sidebar/Utils/formatTopologyProperties.tsx
@@ -1,5 +1,5 @@
 import { Topology } from "../../../../api/types/topology";
-import { FormatProperty } from "../../../TopologyCard/Property";
+import { FormatProperty } from "../../TopologyCard/Property";
 
 type GroupedProperties = {
   type: "GROUPED";
