+++ src/components/Topology/TopologyBreadcrumbs/index.tsx
@@ -1,12 +1,12 @@
-import { Icon } from "../Icon";
-import { useComponentNameQuery } from "../../api/query-hooks";
+import { Icon } from "../../Icon";
+import { useComponentNameQuery } from "../../../api/query-hooks";
 import { useMemo } from "react";
-import { Head } from "../Head/Head";
+import { Head } from "../../Head/Head";
 import {
   BreadcrumbChild,
   BreadcrumbNav,
   BreadcrumbRoot
-} from "../BreadcrumbNav";
+} from "../../BreadcrumbNav";
 
 type TopologyBreadcrumbsProps = {
   topologyId?: string;
