+++ src/components/Topology/TopologyPage/TopologyFilterBar.tsx
@@ -1,11 +1,11 @@
 import { useMemo } from "react";
 import { useSearchParams } from "react-router-dom";
-import { GetTopologyApiResponse } from "../../api/services/topology";
-import { allOption } from "../../pages/TopologyPage";
-import { AgentNamesDropdown } from "../Agents/AgentNamesDropdown";
-import { ComponentLabelsDropdown } from "../Dropdown/ComponentLabelsDropdown";
-import { ComponentTypesDropdown } from "../Dropdown/ComponentTypesDropdown";
-import { ReactSelectDropdown, StateOption } from "../ReactSelectDropdown";
+import { GetTopologyApiResponse } from "../../../api/services/topology";
+import { allOption } from "../../../pages/TopologyPage";
+import { AgentNamesDropdown } from "../../Agents/AgentNamesDropdown";
+import { ComponentLabelsDropdown } from "../../Dropdown/ComponentLabelsDropdown";
+import { ComponentTypesDropdown } from "../../Dropdown/ComponentTypesDropdown";
+import { ReactSelectDropdown, StateOption } from "../../ReactSelectDropdown";
 import TopologyPopOver from "../TopologyPopover";
 import { TopologySort } from "../TopologyPopover/topologySort";
