+++ src/components/Topology/TopologyPopover/topologySort.tsx
@@ -3,11 +3,11 @@ import { uniq } from "lodash";
 import { LegacyRef } from "react";
 import { BsSortDown, BsSortUp } from "react-icons/bs";
 import { useSearchParams } from "react-router-dom";
-import { ValueType } from "../../api/types/common";
-import { Topology } from "../../api/types/topology";
-import { useOnMouseActivity } from "../../hooks/useMouseActivity";
-import { saveSortBy, saveSortOrder } from "../../pages/TopologyPage";
-import { isDate } from "../../utils/date";
+import { ValueType } from "../../../api/types/common";
+import { Topology } from "../../../api/types/topology";
+import { useOnMouseActivity } from "../../../hooks/useMouseActivity";
+import { saveSortBy, saveSortOrder } from "../../../pages/TopologyPage";
+import { isDate } from "../../../utils/date";
 
 const STATUS = {
   info: 4,
